Moved map_colors utility to plotting/util.py
