+++ holoviews/plotting/bokeh/util.py
@@ -53,22 +53,6 @@ def get_cmap(cmap):
     return cm.get_cmap(cmap)
 
 
-def map_colors(arr, crange, cmap):
-    """
-    Maps an array of values to RGB hex strings, given
-    a color range and colormap.
-    """
-    if crange:
-        cmin, cmax = crange
-    else:
-        cmin, cmax = np.nanmin(arr), np.nanmax(arr)
-    arr = (arr - cmin) / (cmax-cmin)
-    arr = np.ma.array(arr, mask=np.logical_not(np.isfinite(arr)))
-    arr = cmap(arr)*255
-    return ["#{0:02x}{1:02x}{2:02x}".format(*(int(v) for v in c[:-1]))
-            for c in arr]
-
-
 def mpl_to_bokeh(properties):
     """
     Utility to process style properties converting any
