+++ holoviews/plotting/util.py
@@ -1,3 +1,4 @@
+import numpy as np
 import param
 
 from ..core import (HoloMap, DynamicMap, CompositeOverlay, Layout,
@@ -239,3 +240,22 @@ def closest_match(match, specs, depth=0):
         else:
             return sorted(match_lengths, key=lambda x: -x[1])[0][0]
 
+
+def map_colors(arr, crange, cmap, hex=True):
+    """
+    Maps an array of values to RGB hex strings, given
+    a color range and colormap.
+    """
+    if crange:
+        cmin, cmax = crange
+    else:
+        cmin, cmax = np.nanmin(arr), np.nanmax(arr)
+    arr = (arr - cmin) / (cmax-cmin)
+    arr = np.ma.array(arr, mask=np.logical_not(np.isfinite(arr)))
+    arr = cmap(arr)
+    if hex:
+        arr *= 255
+        return ["#{0:02x}{1:02x}{2:02x}".format(*(int(v) for v in c[:-1]))
+                for c in arr]
+    else:
+        return arr
